Project Setup Etch-a-sketch
The first commit is related to setting up the project for Etch-a-sketch.
This project is part of The Odin Project curriculum. This first commit
included all HTML links to scripts, stylesheets, and third party
assets such as Google Fonts and Font Awesome. Additionally, it included
adding our favicon to our site as well.